문의사항이미지  아그리고 회원등급 하는중 유석호

+++ hompage/src/main/java/com/example/demo/entity/memberGrade/MemberGrade.java
@@ -8,5 +8,6 @@ import lombok.*;
 public class MemberGrade {
 	private Long gradeCode;
 	private String gradeName;
-	private Integer pointRate;
+	private Double pointRate;		
 }
+
